fix(quiz): resolve post-commit findings on diagram_label (#697)
- code-reviewer BLOCKING: extract RunwayBody to rwy-2709-runway-body.tsx
  (rwy-2709-lh-pattern.tsx 155->102, back under the 150-line component cap)
- semantic-reviewer ISSUE: plan.md integration-test count 356->357 (+45->+46)
  — accounts for the distinct-label reject test the prior fixup added; also
  adds a duplicate-label entry to the CHECK-rejection list
- semantic-reviewer SUGGESTION: seed comments updated for the commit-2 label
  rename (Approach->Go-around, leg labels); comment-only, answer data unchanged
